+++ CHANGES.rst
@@ -17,6 +17,8 @@ Bug fixes:
 - Do not fail on invalid objects in referenec_catalog. Fixes #32 and #48
   [pbauer]
 
+- Fix undeclared hard dependency on Archetypes
+  [tomgross]
 
 3.3.1 (2017-06-01)
 ------------------